// src/types/company.ts
export interface CompanySettings {
  id: string;
  nameAr: string;
  nameEn?: string | null;
  address?: string | null;
  phone1?: string | null;
  phone2?: string | null;
  email?: string | null;
  taxNumber?: string | null;
  logo?: string | null;
  footerText?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}